remove unsed code and duplicate type
