--- backend/src/chat-room/dto/create-chat-room.dto.ts
@@ -1,10 +1,0 @@
-import { IsString, IsOptional, IsNumber } from 'class-validator';
-export class CreateChatRoomDto {
-  @IsOptional()
-  @IsString()
-  name: string;
-
-  @IsOptional()
-  @IsNumber()
-  created_by: number;
-}
